move code to responsible classes
